+++ test/get_arch_application_test.dart
@@ -10,6 +10,14 @@ import 'package:get_arch_core/profile/get_arch_application.dart';
 main() {
   test('show printPackageConfigInfo', () {
     GetArchCorePackage().init(
-        EnvConfig('Test GetArch', '0.0.1', DateTime(2020, 6, 20), 'dev'), true);
+        EnvConfig('Test GetArch', '0.0.1', DateTime(2020, 6, 20), EnvSign.dev),
+        true);
+  });
+
+  test('show EnvSign value', () {
+    expect(EnvSign.dev, 'dev');
+  });
+  test('get EnvSign from String ', () {
+    expect(EnvSign.fromString('dev'), EnvSign.dev);
   });
 }